+++ packages/engine-multi/CHANGELOG.md
@@ -1,5 +1,12 @@
 # engine-multi
 
+## 1.6.14
+
+### Patch Changes
+
+- Updated dependencies [cb97744]
+  - @openfn/compiler@1.1.5
+
 ## 1.6.13
 
 ### Patch Changes